+++ CoreArrangementPlugins/src/au/gov/asd/tac/constellation/plugins/arrangements/GraphTaxonomyArranger.java
@@ -152,7 +152,7 @@ public abstract class GraphTaxonomyArranger implements Arranger {
             final int steps = taxa.size() + 1;
             int step = 0;
             for (final Map.Entry<Integer, Set<Integer>> entry : taxa.entrySet()) {
-                if (taxonomy.getArrangeRectangularly(entry.getKey())) {
+                if (taxonomy.isArrangeRectangularly(entry.getKey())) {
                     if (interaction != null) {
                         interaction.setProgress(step, steps, "Arrange grid...", true);
                     }