+++ CoreGraphUtilities/src/au/gov/asd/tac/constellation/graph/utilities/hashmod/HashmodPanel.java
@@ -108,15 +108,15 @@ public class HashmodPanel extends javax.swing.JPanel {
         }
     }
 
-    public boolean getCreateVertexes() {
+    public boolean isCreateVertexesSelected() {
         return createAllCheckbox.isSelected();
     }
 
-    public boolean getCreateAttributes() {
+    public boolean isCreateAttributesSelected() {
         return createAttributesCheckbox.isSelected();
     }
 
-    public boolean getCreateTransactions() {
+    public boolean isCreateTransactionsSelected() {
         return createTransactionsCheckbox.isSelected();
     }